Improved UI and API.
