+++ SecclBackend/Controllers/PortfolioController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using SecclBackend.Services;
-using SecclBackend.Models;
+using SecclShared.Models;
 
 namespace SecclBackend.Controllers
 {
@@ -31,13 +31,13 @@ namespace SecclBackend.Controllers
 
             if (portfolioData == null || portfolioData.Count == 0)
             {
-                return NotFound("Portfolio data not found for the specified client.");
+                return NotFound(new ErrorResponse { Message = "Portfolio data not found for the specified client.", StatusCode = 404 });
             }
 
             return Ok(portfolioData.First()); // Return the first portfolio if only one is expected
         }
 
-        //Get aggregated total value of three portfolios
+        //Get aggregated total value of multiple portfolios
         [HttpGet("aggregated-total")]
         public async Task<IActionResult> GetAggregatedTotalValue([FromQuery] string clientIds)
         {
@@ -48,7 +48,7 @@ namespace SecclBackend.Controllers
 
             if (clientIdList == null || clientIdList.Count == 0)
             {
-                return BadRequest("Client IDs are required.");
+                return BadRequest(new ErrorResponse { Message = "Client IDs are required.", StatusCode = 400 });
             }
 
             var token = await _secclApiService.GetAccessTokenAsync();
@@ -74,7 +74,7 @@ namespace SecclBackend.Controllers
 
             if (clientIdList == null || clientIdList.Count == 0)
             {
-                return BadRequest("Client IDs are required.");
+                return BadRequest(new ErrorResponse { Message = "Client IDs are required.", StatusCode = 400 });
             }
 
             var token = await _secclApiService.GetAccessTokenAsync();
@@ -98,6 +98,37 @@ namespace SecclBackend.Controllers
             }
 
             return Ok(accountTypeTotals);
+        }
+
+        [HttpPost("batch-aggregated-data")]
+        public async Task<IActionResult> GetBatchAggregatedData([FromBody] List<string> clientIds)
+        {
+            if (clientIds == null || !clientIds.Any())
+            {
+                return BadRequest(new SecclShared.Models.ErrorResponse { Message = "Client IDs are required.", StatusCode = 400 });
+            }
+
+            var token = await _secclApiService.GetAccessTokenAsync();
+            decimal totalValue = 0;
+            var accountTypeTotals = new Dictionary<string, decimal>();
+
+            foreach (var clientId in clientIds)
+            {
+                var portfolios = await _secclApiService.FetchPortfolioData(token, clientId);
+                totalValue += portfolios.Sum(p => p.CurrentValue);
+                foreach (var portfolio in portfolios)
+                {
+                    if (accountTypeTotals.ContainsKey(portfolio.AccountType))
+                        accountTypeTotals[portfolio.AccountType] += portfolio.CurrentValue;
+                    else
+                        accountTypeTotals[portfolio.AccountType] = portfolio.CurrentValue;
+                }
+            }
+
+            return Ok(new {
+                TotalValue = totalValue,
+                AggregatedTotals = accountTypeTotals
+            });
         }
     }
 }
