+++ SecclBackend/Models/ErrorResponse.cs
@@ -1,0 +1,1 @@
+// File moved to SecclShared. Delete this file.
