+++ SecclBackend/SecclBackend.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SecclShared\SecclShared.csproj" />
+  </ItemGroup>
+
 </Project>
