+++ SecclBackend/appsettings.json
@@ -5,5 +5,11 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "SecclApi": {
+    "BaseUrl": "https://pfolio-api-staging.seccl.tech",
+    "FirmId": "P1IMX",
+    "Username": "nelahi6642@4tmail.net",
+    "Password": "DemoBDM1"
+  }
 }
