+++ SecclShared/ErrorResponse.cs
@@ -1,0 +1,8 @@
+namespace SecclShared.Models
+{
+    public class ErrorResponse
+    {
+        public string Message { get; set; }
+        public int StatusCode { get; set; }
+    }
+}
